requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic>=2.9.0
pydantic-settings==2.1.0
python-multipart==0.0.6
google-genai>=1.59.0
python-dotenv==1.0.0
